+++ build.gradle
@@ -11,24 +11,27 @@ plugins {
 
 // Repository Source Configuration
 repositories {
-  // Using the maven central repository source
-  mavenCentral()
+  mavenCentral() // Using the maven central repository source
+}
+
+dependencies {
+    testImplementation 'io.rest-assured:rest-assured:5.3.1' // add rest-assured dependencies
+    testImplementation 'org.testng:testng:7.8.0' // Add test framework TestNG dependency
+    implementation 'org.uncommons:reportng:1.1.4' // testng report dependency
+    implementation 'org.slf4j:slf4j-api:2.0.9' // Add test log dependency
+    implementation 'org.slf4j:slf4j-simple:2.0.9' // Add test log dependency
+    implementation group: 'com.google.inject', name: 'guice', version: '7.0.0'
 }
 
 // test config
 test {
-    useTestNG() // Tell Gradle to use TestNG as a testing framework.
+    reports.html.required = false // 禁用 HTML 报告生成
+    reports.junitXml.required = false // 禁用 JUnit XML 报告生成
+    // Tell Gradle to use TestNG as a testing framework.
+    useTestNG() {
+        useDefaultListeners = true
+        suites 'src/test/resources/testng.xml' // Specify the path to the test suite XML file
+    }
     testLogging.showStandardStreams = true // Output test logs to the console
     testLogging.events "passed", "skipped", "failed"
-}
-
-// testng framework config
-test.useTestNG() {
-    suites "src/test/resources/testng.xml" // Specify the path to the test suite XML file
-}
-dependencies {
-    testImplementation 'io.rest-assured:rest-assured:5.3.2' // add rest-assured dependencies
-    testImplementation 'org.testng:testng:7.8.0' // Add test framework TestNG dependency
-    testImplementation 'org.slf4j:slf4j-api:2.0.9' // Add test log dependency
-    testImplementation 'org.slf4j:slf4j-simple:2.0.9' // Add test log dependency
 }
