+++ src/test/resources/testng.xml
@@ -1,9 +1,9 @@
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
-<suite name="restassured-gradleTestSuite">
-    <test name="Demo">
-        <classes>
-            <class name="com.example.GetDemoAPITest"/>
-            <!-- add more test -->
-        </classes>
-    </test>
-</suite>
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
+<suite name="restAssured-gradleTestSuite">
+<test thread-count="1" name="Demo">
+    <classes>
+        <class name="com.example.TestDemo"/>
+    </classes>
+</test> <!-- Test -->
+</suite> <!-- Suite -->
